@@ -1,26 +1,51 @@
-const _ = require('lodash');
-const request = require("request");
-const url = require('url');
-const EventEmitter = require('events').EventEmitter;
+import request from 'request';
+import url from 'url';
+import { EventEmitter } from 'node:events';
+import { newLogger } from '../../common/logger.js';
+import Utils from '../../common/utils.js';
 
-const config = require('config');
-const Logger = require("./logger.js");
-const Utils = require("./utils.js");
+const Logger = newLogger('callback-emitter');
+
+// A simple string that identifies the event
+const simplifiedEvent = (event) => {
+  if (event.event != null) {
+    event = event.event
+  }
+  try {
+    const eventJs = JSON.parse(event);
+    return `event: { name: ${(eventJs.data != null ? eventJs.data.id : undefined)}, timestamp: ${(eventJs.data.event != null ? eventJs.data.event.ts : undefined)} }`;
+  } catch (e) {
+    return `event: ${event}`;
+  }
+};
 
 // Use to perform a callback. Will try several times until the callback is
 // properly emitted and stop when successful (or after a given number of tries).
 // Used to emit a single callback. Destroy it and create a new class for a new callback.
 // Emits "success" on success, "failure" on error and "stopped" when gave up trying
 // to perform the callback.
-module.exports = class CallbackEmitter extends EventEmitter {
-
-  constructor(callbackURL, message, permanent) {
+export default class CallbackEmitter extends EventEmitter {
+  constructor(callbackURL, event, permanent, options = {}) {
     super();
     this.callbackURL = callbackURL;
-    this.message = message;
+    this.event = event;
+    this.message = JSON.stringify(event);
     this.nextInterval = 0;
     this.timestamp = 0;
     this.permanent = permanent;
+
+    this._permanentIntervalReset = options.permanentIntervalReset || 8;
+    this._serverDomain = options.domain;
+    this._secret = options.secret;
+    this._bearerAuth = options.auth2_0;
+    this._requestTimeout = options.requestTimeout;
+    this._retryIntervals = options.retryIntervals || [
+      1000,
+      2000,
+      5000,
+      10000,
+      30000,
+    ];
   }
 
   start() {
@@ -38,20 +63,20 @@ module.exports = class CallbackEmitter extends EventEmitter {
           this.emit("failure", error);
 
           // get the next interval we have to wait and schedule a new try
-          const interval = config.get("hooks.retryIntervals")[this.nextInterval];
+          const interval = this.retryIntervals[this.nextInterval];
           if (interval != null) {
-            Logger.warn(`[Emitter] trying the callback again in ${interval/1000.0} secs`);
+            Logger.warn(`trying the callback again in ${interval/1000.0} secs`);
             this.nextInterval++;
             this._scheduleNext(interval);
 
           // no intervals anymore, time to give up
           } else {
-            this.nextInterval = config.get("hooks.permanentIntervalReset"); // Reset interval to permanent hooks
+            this.nextInterval = this._permanentIntervalReset;
             if(this.permanent){
               this._scheduleNext(this.nextInterval);
             }
             else {
-              return this.emit("stopped");
+              this.emit("stopped");
             }
           }
         }
@@ -62,10 +87,10 @@ module.exports = class CallbackEmitter extends EventEmitter {
 
   _emitMessage(callback) {
     let data, requestOptions;
-    const serverDomain = config.get("bbb.serverDomain");
-    const sharedSecret = config.get("bbb.sharedSecret");
-    const bearerAuth = config.get("bbb.auth2_0");
-    const timeout = config.get('hooks.requestTimeout');
+    const serverDomain = this._serverDomain;
+    const sharedSecret = this._secret;
+    const bearerAuth = this._bearerAuth;
+    const timeout = this._requestTimeout;
 
     // data to be sent
     // note: keep keys in alphabetical order
@@ -97,7 +122,7 @@ module.exports = class CallbackEmitter extends EventEmitter {
       // get the final callback URL, including the checksum
       const urlObj = url.parse(this.callbackURL, true);
       let callbackURL = this.callbackURL;
-      callbackURL += _.isEmpty(urlObj.search) ? "?" : "&";
+      callbackURL += Utils.isEmpty(urlObj.search) ? "?" : "&";
       callbackURL += `checksum=${checksum}`;
 
       requestOptions = {
@@ -116,27 +141,14 @@ module.exports = class CallbackEmitter extends EventEmitter {
         return !((statusCode >= 200 && statusCode < 300) || statusCode == 401)
     };
 
-    request(requestOptions, function(error, response, body) {
+    request(requestOptions, (error, response) => {
       if ((error != null) || responseFailed(response)) {
-        Logger.warn(`[Emitter] error in the callback call to: [${requestOptions.uri}] for ${simplifiedEvent(data)} error: ${error} status: ${response != null ? response.statusCode : undefined}`);
+        Logger.warn(`error in the callback call to: [${requestOptions.uri}] for ${simplifiedEvent(data)} error: ${error} status: ${response != null ? response.statusCode : undefined}`);
         callback(error, false);
       } else {
-        Logger.info(`[Emitter] successful callback call to: [${requestOptions.uri}] for ${simplifiedEvent(data)}`);
+        Logger.info(`successful callback call to: [${requestOptions.uri}] for ${simplifiedEvent(data)}`);
         callback(null, true);
       }
     });
   }
-};
-
-// A simple string that identifies the event
-var simplifiedEvent = function(event) {
-  if (event.event != null) {
-    event = event.event
-  }
-  try {
-    const eventJs = JSON.parse(event);
-    return `event: { name: ${(eventJs.data != null ? eventJs.data.id : undefined)}, timestamp: ${(eventJs.data.event != null ? eventJs.data.event.ts : undefined)} }`;
-  } catch (e) {
-    return `event: ${event}`;
-  }
-};
+}
